completed the current temp section
