+++ src/views/base.js
@@ -1,0 +1,9 @@
+export const dom = {
+  searchForm: document.getElementsByTagName('form'),
+  cityInput: document.querySelector('.search__city'),
+  countryInput: document.querySelector('.search__country'),
+  searchBtn: document.querySelector('.search__btn'),
+  infoSection: document.querySelector('.info-section'),
+  daySection: document.querySelector('.day-section'),
+}
+
